Change the signature of merge_sort()

+++ src/core/util.h
@@ -68,7 +68,7 @@ void* __reallocx(void *ptr, size_t bytes, const char* location);
 /* General utilities */
 int game_version_compare(int sup_version, int sub_version, int wip_version); /* compare to this version of the game engine */
 void fatal_error(const char *fmt, ...); /* crash the program with a message */
-void merge_sort(void *base, size_t num, size_t size, int (*comparator)(const void*,const void*)); /* similar to stdlib's qsort, but merge_sort is a stable sorting algorithm */
+void merge_sort(void *base, int num, size_t size, int (*comparator)(const void*,const void*)); /* similar to stdlib's qsort, but merge_sort is a stable sorting algorithm */
 float lerp(float a, float b, float t); /* linear interpolation */
 float lerp_angle(float alpha, float beta, float t); /* alpha, beta in radians */
 uint64_t random64(); /* pseudo-random 64-bit number */
